fix: race conditions, sheet bloat, and code duplication
- Replace getChatData→saveChat with patchChatData in
  useGenerationErrorHandler to prevent race conditions (Rule 3)
- Extract gallery logic from CharacterCardSheet into
  useCharacterGallery composable (Sheet trap rule)
- Deduplicate fileToDataUrl/blobToDataUrl into shared imageUtils
- Document abort-path clearGenerationState delegation in
  generation.md and INVARIANTS.md
